chore: ajustes de estilo e correcao login

+++ src/app/context/AuthContext.tsx
@@ -2,12 +2,12 @@ import React, { createContext, useState, useContext } from "react";
 
 const AuthContext = createContext<{
   user: User | null;
-  loginEfetuado: (userId: string, token: string) => void;
+  loginEfetuado: (userId: string) => void;
   logout: () => void;
   getUser: () => User | null;
 }>({
   user: null,
-  loginEfetuado: (userId: string, token: string) => {},
+  loginEfetuado: (userId: string) => {},
   logout: () => {},
   getUser: () => null,
 });
@@ -23,17 +23,19 @@ interface AuthProviderProps {
 export function AuthProvider({ children }: AuthProviderProps) {
   const [user, setUser] = useState<User | null>(null);
   const auth = getAuth();
-  const userFirebase = auth.currentUser;
+  
 
-  const loginEfetuado = async (userId: string, token: string) => {
+  const loginEfetuado = async (userId: string) => {
+    
+    console.log('userId', userId);
     try {
       const response = await fetch(
         `https://api-dkviu3xq7a-uc.a.run.app/user?userId=${userId}`
       );
       const userData = await response.json();
-      console.log("response", response.status);
       if (response.status === 404) {
-        excluirUsuarioAuth();
+        console.log('excluir', userId);
+        excluirUsuarioAuth(userId);
       }
       setUser(userData);
     } catch (error: any) {
@@ -41,18 +43,29 @@ export function AuthProvider({ children }: AuthProviderProps) {
     }
   };
 
-  const excluirUsuarioAuth = async () => {
-    try {
-      if (userFirebase) {
-        await deleteUser(userFirebase);
+  const excluirUsuarioAuth = async (id: string) => {
+    const userFirebase = auth.currentUser;
+    console.log('userFirebase', userFirebase);
+    if(id === userFirebase?.uid){
+      try {
+        if (userFirebase) {
+          await deleteUser(userFirebase);
+          logout();
+        }
+      } catch (error) {
+        console.error("Erro ao excluir usuário:", error);
       }
-    } catch (error) {
-      console.error("Erro ao excluir usuário:", error);
     }
   };
 
-  const logout = () => {
-    setUser(null);
+  const logout = async () => {
+    try {
+      await auth.signOut();
+      setUser(null);
+    } catch (e) {
+        console.error(e);
+    }
+    
   };
 
   const getUser = () => {
